about and contact page (#4)
* UI Bug Fix
bugfix: changes the burger positioning to be absolute to the navbar. when viewing on mobile, the burger menu was shifted away from the window edge. this was not replicatable when viewing in the browser mobile view during development.
bugfix: hides the overflow on the inner container which created a scrollbar that caused the burger menu to shift when viewing on desktop.
refactor: fixes relative paths on portfolio.

* add about and contact pages
feat: adds a fully styled and responsive about us page.
feat: adds a fully styled and responsive contact page.
feat: adds contact page form validation.
feat: adds contact page map feature.
refactor: changes px to rem.
refactor: changes font size to default.